feat: Introduce a `Quantity` and `Unit` system for resources and templates, and refactor the application entry point.

+++ internal/adapters/inbound/rest/templates/dto.go
@@ -7,16 +7,16 @@ type StepJSONResponseDTO struct {
 }
 
 type ResourceRefResponseDTO struct {
-	ResourceID int    `json:"resource_id"`
-	Quantity   int    `json:"quantity"`
-	Unit       string `json:"unit"`
+	ResourceID int     `json:"resource_id"`
+	Quantity   float64 `json:"quantity"`
+	Unit       string  `json:"unit"`
 }
 
 type TemplateJSONResponseDTO struct {
 	ID          int                      `json:"id"`
 	Name        string                   `json:"name"`
 	Description string                   `json:"description"`
-	Quantity    int                      `json:"quantity"`
+	Quantity    float64                  `json:"quantity"`
 	Unit        string                   `json:"unit"`
 	Difficulty  int                      `json:"difficulty"`
 	Steps       []StepJSONResponseDTO    `json:"steps"`
@@ -32,7 +32,7 @@ type TemplateListJSONResponseDTO struct {
 type CreateTemplateRequestDTO struct {
 	Name        string           `json:"name"`
 	Description string           `json:"description"`
-	Quantity    int              `json:"quantity"`
+	Quantity    float64          `json:"quantity"`
 	Unit        string           `json:"unit"`
 	Difficulty  int              `json:"difficulty"`
 	Steps       []StepDTO        `json:"steps"`
@@ -43,7 +43,7 @@ type UpdateTemplateRequestDTO struct {
 	ID          int              `json:"id"`
 	Name        string           `json:"name"`
 	Description string           `json:"description"`
-	Quantity    int              `json:"quantity"`
+	Quantity    float64          `json:"quantity"`
 	Unit        string           `json:"unit"`
 	Difficulty  int              `json:"difficulty"`
 	Steps       []StepDTO        `json:"steps"`
@@ -51,9 +51,9 @@ type UpdateTemplateRequestDTO struct {
 }
 
 type ResourceRefDTO struct {
-	ResourceID int    `json:"resource_id"`
-	Quantity   int    `json:"quantity"`
-	Unit       string `json:"unit"`
+	ResourceID int     `json:"resource_id"`
+	Quantity   float64 `json:"quantity"`
+	Unit       string  `json:"unit"`
 }
 
 type StepDTO struct {
